Feat: "웹소켓 서버 구조 픽스"

+++ websocket/travel/eventHandler.js
@@ -1,5 +1,7 @@
 import { eventEmitter, errorEmitter, CAST_TYPES } from "./emitter.js"
-import { join, updateStaytime, swapSchedule, createSchedule, deleteSchedule } from "./stateManager.js"
+import { checkAuthorized, join, revokeAllAuthorities } from "./stateManager.js"
+import { updateStaytime, swapSchedule, createSchedule, deleteSchedule } from "./stateManager.js"
+import { grantAuthority, revokeAuthority } from "./stateManager.js"
 
 // ERROR
 const ERRORS = {
@@ -40,13 +42,30 @@ const EVENTS = {
     callback: (socket, namespace, room, roomTable, eventName, { id }) => {
       try {
         join(room, roomTable, id)
+        socket.data.id = id
         eventEmitter({ socket, namespace, room }, 
           CAST_TYPES.BROADCAST_SERVER, eventName, { id })
       }
       catch (err) {
         errorEmitter({ socket, namespace, room },
-          CAST_TYPES.UNICAST, ERRORS.JOIN_ERROR)
+          CAST_TYPES.UNICAST, ERRORS.INPUT_ERROR(err))
         socket.disconnect()
+      }
+    }
+  },
+  LEAVE_EVENT: {
+    eventName: "leave",
+    callback: (socket, namespace, room, roomTable, eventName, _) => {
+      const id = socket.data.id
+      try {
+        leave(room, roomTable, id)
+        revokeAllAuthorities(room, roomTable, { id })
+        eventEmitter({ socket, namespace, room }, 
+          CAST_TYPES.BROADCAST_SERVER, eventName, { id })
+      }
+      catch (err) {
+        errorEmitter({ socket, namespace, room },
+          CAST_TYPES.UNICAST, ERRORS.INPUT_ERROR(err))
       }
     }
   },
@@ -68,15 +87,53 @@ const EVENTS = {
         })
     }
   },
+  GRANT_AUTHORITY_EVENT: {
+    eventName: "grant authority",
+    callback: (socket, namespace, room, roomTable, eventName, 
+        { id, authorityName, day }) => {
+      try {
+        grantAuthority(room, roomTable, { id, authorityName, day })
+        eventEmitter({ socket, namespace, room },
+          CAST_TYPES.UNICAST, eventName, {
+            granted: 1
+          })
+      }
+      catch (err) {
+        eventEmitter({ socket, namespace, room },
+          CAST_TYPES.UNICAST, eventName, {
+            granted: 0
+          })
+      }
+    }
+  },
+  REVOKE_AUTHORITY_EVENT: {
+    eventName: "revoke authority",
+    callback: (socket, namespace, room, roomTable, eventName, 
+        { id, authorityName, day }) => {
+      try {
+        revokeAuthority(room, roomTable, { id, authorityName, day })
+        eventEmitter({ socket, namespace, room },
+          CAST_TYPES.UNICAST, eventName, {
+            revoked: 1
+          })
+      }
+      catch (err) {
+        eventEmitter({ socket, namespace, room },
+          CAST_TYPES.UNICAST, eventName, {
+            revoked: 0
+          })
+      }
+    }
+  },
   UPDATE_STAYTIME_EVENT: {
     eventName: "update staytime",
     callback: (socket, namespace, room, roomTable, eventName, 
-        { day, turn, staytime }) => {
+        { day, turn, stayTime }) => {
       try {
-        updateStaytime(room, roomTable, { day, turn, staytime })
+        updateStaytime(room, roomTable, { day, turn, stayTime })
         eventEmitter({ socket, namespace, room },
           CAST_TYPES.BROADCAST_SERVER, eventName, 
-          { day, turn, staytime })
+          { day, turn, stayTime })
       }
       catch (err) {
         errorEmitter({ socket, namespace, room },
@@ -89,6 +146,11 @@ const EVENTS = {
     callback: (socket, namespace, room, roomTable, eventName, 
         { day, turn1, turn2 }) => {
       try {
+        checkAuthorized(room, roomTable, { 
+          id: socket.data.id,
+          authorityName: "schedules",
+          day
+         })
         swapSchedule(room, roomTable, { day, turn1, turn2 })
         eventEmitter({ socket, namespace, room },
           CAST_TYPES.BROADCAST_SERVER, eventName,
@@ -105,6 +167,11 @@ const EVENTS = {
     callback: (socket, namespace, room, roomTable, eventName,
         { day, placeUid, placeName, lat, lng }) => {
       try {
+        checkAuthorized(room, roomTable, {
+          id: socket.data.id,
+          authorityName: "schedules",
+          day
+        })
         createSchedule(room, roomTable, { day, placeUid, placeName, lat, lng })
         eventEmitter({ socket, namespace, room },
           CAST_TYPES.BROADCAST_SERVER, eventName,
@@ -121,6 +188,11 @@ const EVENTS = {
     callback: (socket, namespace, room, roomTable, eventName,
         { day, turn }) => {
       try {
+        checkAuthorized(room, roomTable, {
+          id: socket.data.id,
+          authorityName: "schedules",
+          day
+        })
         deleteSchedule(room, roomTable, { day, turn })
         eventEmitter({ socket, namespace, room },
           CAST_TYPES.BROADCAST_SERVER, eventName,
